refactor(LottoGenerator) : LottoGenerator Domain으로 영역 변경
