+++ src/test/java/lotto/domain/LottosTest.java
@@ -6,8 +6,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
-import lotto.controller.generator.LottoAutoGenerator;
-import lotto.controller.generator.LottoManualGenerator;
+import lotto.domain.generator.LottoAutoGenerator;
+import lotto.domain.generator.LottoManualGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
